twitter link preview
in server.js inside the * route we check if the request came from a campaigns folder, get information for the link preview from the db and replace it in the og meta tags in the index.html.
in order to get that call we rename the index.html into _index.html so the express static server does not find the file and the star route gets the call
index.html is adjusted to include og tags.
react script for build command is adjusted to move the file from index.html to _index.html